style(RacingResult) : 최대 이동 거리를 활용한 최종 우승자 탐색 메서드 수정

+++ src/main/java/racingcar/model/RacingResult.java
@@ -20,7 +20,7 @@ public class RacingResult {
 
     public List<String> findWinner() {
         return racingRecords.stream()
-                .filter(racingCar -> racingCar.getDistance() == maxDistance)
+                .filter(racingCar -> racingCar.getDistance() == maxDistance.intValue())
                 .map(RacingCar::getName)
                 .toList();
     }
